Fix regression for username fallback
While usernames were now picked up correctly as a fallback from the
filename if none were present in the fields themselves, if there were
any in the fields the program would return an empty string.

This fixes it for both cases, the username existing in the entry fields
or only in the filename. Entry fields will still take precedence.